Second list of changes on the backend. The following files have been changed:
1. addtripmember.php
2. deletetripmembers.php
3. invitetotrip.php
4. admin.php.php

addtripmember.php and deletetripmembers.php have been updated to remove any
forms or FBML commands. Specifically, both these expect to get some
information from the front-end, modify the MYSQL database accordingly and
return a json object. The returning json object will notify the front-end of
any error and the front-end can respond accordingly.

invitetotrip.php has also been modified to remove any FBML calls. This means
the existing multi-friend-selector call and similar calls have been removed.
Now, invite to trip works in two ways:

a) When the tripid and facebook userid are passed from the front-end, the
notifications table is checked, and if ok, returns a json object with the
already invited friends as well as trip related information

b) When the tripid, facebook userid and the list of invited friends' userid's
are passed, invitetotrip updates the notification table. Also, the tripmembers
table is updated to reflect that these people have been invited

admin.php - also is being modified. The modifications are not complete for
this file. The first change is to move away from
hard-coded userid's for administrative access. To this end, this file now
reads the admin users from a "cmcadmins" table. This table does not exist yet.