In this commit, the search functionality is integrated on the backend. Previous version had three different searches - basic, advanced and zipcode search. This version does not differentiate between different searches - they are all in one. For this purpose, the functionality is integrated into a new function called searchresults.php.
For proper working, the facebook userid (fbid) and a searchkeys need to be
passed as GET or POST parameters from the front-end. The searchkeys needs to
be json_encoded, base64_encoded object. On the backend, the object is first
base64_decoded, then json_decoded. This new object is then used for processing
the search results. 

The search terms currently can be:

1. name - for the name of the person
2. relg - religion id
3. medskills - medical skills id
4. otherskills - other skills id
5. spiritserv - spiritual skills id
6. country - country id to be searches
7. region - region id
8. dur - trip duration id
9. z - zip code array. This array needs to have two values: the first value is
the zipcode, and the search value is the searchradius.


If no error is encountered, the search results into a json encoded string with
relevant facebook ids.

If the zipcode (searchkeys->{'z'}) is not specified, no sorting is done. The
resulting facebook ids are listed according to the result of the query. If the
zipcode is provided, the results are filtered according to the search
radius,and the closest results are listed first.


The following files are removed since they are no longer needed:
searchzipresults.php