This is a very solid version of the search code as it is now. Search is still not feature complete.
This commit first and foremost fixes several bugs:
 * [99620871] IE only: multi-parameter searching breaks pictures
 * [99598451] When moving from final search page to previous page, pictures get intermittently deleted
 * [99598560] Crash: Typing a backslash in to the search box locks up the app
 * [97216525] Quickly clicking back/forward search page buttons [...] renders results mixed with previous page

The work in this commit is all stabilization -- I haven't added new features in this version.

Also, as part of this commit, I have added vim modelines to the files in order to enforce style guidelines.

I have also added a Makefile to the project, and written some custom commands that allow debug code to be removed
quickly from production versions of our Javascript. Run 'make ship' in a branch to make a quicker, smaller version
of the Javascript code (which will be written as cmc.min.js).