The following additional details have been accomplished in this commit:
1. Profile display CSS file
2. Some minor bug fixes to the multiple selection of the profile entry form
without holding the cntrl-key
3. Profile display by retrieving profile information by an AJAX call to the
backend. This functionality is NOT complete - it is still a work in progress.
I'm commiting it so that you can see how it looks so far. Currently, the profile
information retrieval is hard-coded for a specific facebook userid. In the
final application, this information will automatically come from the
FB.getLoginStatus or FB.login calls. The next main step is to figure out a way
to get this information. 
4. Some JQuery bare-bones functionality to associate the "Edit Profile",
"Join Trip" and other buttoms from the profile page. This part is not done
either - additional details need to be added.