Modifications:
1. changes to cmc.js
	a) Added functions to get future CMC trips
	b) Added functionality for updating the trips tab to reflect upcoming trips.
	If no upcoming trips exist, the users are prompted to create a trip when the
	"Trips" tab is clicked
	c) The correct button and click events have been created. Code needs to go
	into these to effectively respond to the callbacks

2. Changed index.php to reflect the changes in cmc.js - Basically modifed the
Trips tab information

3. Changes to the backend php files to make sure that the ajax calls produce
appropriate behavior. One thing noticed was that when there is a warning on
the php side, JQUERY ajax calls don't go into the success part of the ajax
call. One reason for the warnings was that the time zone was not set and date
functions were being used. So, updated common.php to set the default time zone
to the US East coast