One Major change:
Changed the event handler so that the "live" feature is not used. Now, the
user can click any trip description and the corresponding trip profile
information is displayed.

There was a problem with displaying wrong trip information when clicking from
the upcoming (future) trips button. Fixed that.

Several small changes to cmc.js and index.php

Also some changes to the backend php files.