This change fixes the following bugs:
-Long trip profiles render off the end of the page

-The text of the trip profile is too close to the profile picture

I fixed these by removing specific CSS styling elements for the div tags on the profile and trip tabs, and combined them into one entity class which 
I then reapplied to both. Because the CSS already had the overflow style set, this automatically fixed the trip profile. To fix the spacing, I 
changed the styling for the profile picture on the trip tab to be the same as that of the profile tab. The profile tabs picture styling already had 
the correct spacing.